moved types into the separate file types.ts
