+++ src/components/TodoInfo/TodoIndo.types.ts
@@ -1,4 +1,4 @@
-import { TodoWithUser } from "../../react-app-env";
+import { TodoWithUser } from '../../types';
 
 export type Props = {
   todo: TodoWithUser
