+++ src/components/TodoList/TodoList.types.ts
@@ -1,4 +1,4 @@
-import { TodoWithUser } from '../../react-app-env';
+import { TodoWithUser } from '../../types';
 
 export type Props = {
   todos: TodoWithUser[]
