+++ src/react-app-env.d.ts
@@ -1,18 +1,1 @@
 /// <reference types="react-scripts" />
-
-export interface Todo {
-  id: number,
-  userId: number,
-  title: string,
-  completed: boolean,
-}
-
-export interface User {
-  id: number,
-  name: string,
-  email: string,
-}
-
-export interface TodoWithUser extends Todo {
-  user: User | null,
-}
